+++ src/app/home/_components/letter-count-box.tsx
