+++ src/components/common/index.ts
@@ -1,7 +1,7 @@
+export * from "../../app/home/_components/letter-count-box"
 export * from "./header/header"
 export * from "./header/header-back"
 export * from "./icon"
-export * from "./letter-count-box"
 export * from "./text"
 export * from "./tooltip"
